Refactor Array/Mesh classes
